Move base64 to a feature
Starting work on tests
