+++ scripts/test.sh
@@ -3,4 +3,4 @@
 set -eu
 
 cargo test --all "$@"
-cargo test --all "$@" --features="default serialize"
+cargo test --all "$@" --features="default serialize base64_format"
