+++ src/api/propagation/mod.rs
@@ -160,6 +160,8 @@
 use crate::api;
 use std::collections::HashMap;
 
+#[cfg(feature="base64")]
+pub mod base64_format;
 pub mod binary_propagator;
 pub mod noop;
 pub mod text_propagator;
